// src/types/partials/CodeableConcept.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCoding } from '../partials/Coding';

export type TCodeableConcept = {
  id?: string;
  extension?: TExtension[];
  coding?: TCoding[];
  text?: string;
};