// src/types/partials/Quantity.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TDecimal } from '../simpleTypes/Decimal';
import { TUri } from '../simpleTypes/Uri';

export type TQuantity = {
  id?: string;
  extension?: TExtension[];
  value?: TDecimal;
  comparator?: string;
  unit?: string;
  system?: TUri;
  code?: string;
};