// src/types/partials/Ratio.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TQuantity } from '../partials/Quantity';

export type TRatio = {
  id?: string;
  extension?: TExtension[];
  numerator?: TQuantity;
  denominator?: TQuantity;
};