// src/types/partials/Identifier.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TUri } from '../simpleTypes/Uri';
import { TPeriod } from '../partials/Period';
import { TReference } from '../partials/Reference';

export type TIdentifier = {
  id?: string;
  extension?: TExtension[];
  use?: string;
  type?: TCodeableConcept;
  system?: TUri;
  value?: string;
  period?: TPeriod;
  assigner?: TReference;
};