// src/types/partials/ContactPoint.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TInt } from '../simpleTypes/Int';
import { TPeriod } from '../partials/Period';

export type TContactPoint = {
  id?: string;
  extension?: TExtension[];
  system?: string;
  value?: string;
  use?: string;
  rank?: TInt;
  period?: TPeriod;
};