// src/types/partials/Extension.ts
export type TExtension = {
  id?: string;
  extension?: TExtension[];
  url?: TUri;
  valueBase64Binary?: TBase64Binary;
  valueBoolean?: boolean;
  valueCanonical?: TCanonical;
  valueCode?: string;
  valueDate?: TDate;
  valueDateTime?: TDateTime;
  valueDecimal?: TDecimal;
  valueId?: TId;
  valueInstant?: TInstant;
  valueInteger?: TInt;
  valueMarkdown?: TMarkdown;
  valueOid?: TOid;
  valuePositiveInt?: TInt;
  valueString?: string;
  valueTime?: TTime;
  valueUnsignedInt?: TUnsignedInt;
  valueUri?: TUri;
  valueUrl?: TUrl;
  valueUuid?: TUuid;
  valueAddress?: TAddress;
  valueAge?: TQuantity;
  valueAnnotation?: TAnnotation;
  valueAttachment?: TAttachment;
  valueCodeableConcept?: TCodeableConcept;
  valueCodeableReference?: TCodeableReference;
  valueCoding?: TCoding;
  valueContactPoint?: TContactPoint;
  valueCount?: TQuantity;
  valueDistance?: TQuantity;
  valueDuration?: TQuantity;
  valueHumanName?: THumanName;
  valueIdentifier?: TIdentifier;
  valueMoney?: TMoney;
  valuePeriod?: TPeriod;
  valueQuantity?: TQuantity;
  valueRange?: TRange;
  valueRatio?: TRatio;
  valueRatioRange?: TRatioRange;
  valueReference?: TReference;
  valueSampledData?: TSampledData;
  valueSignature?: TSignature;
  valueTiming?: TTiming;
  valueContactDetail?: TContactDetail;
  valueContributor?: TContributor;
  valueDataRequirement?: TDataRequirement;
  valueExpression?: TExpression;
  valueParameterDefinition?: TParameterDefinition;
  valueRelatedArtifact?: TRelatedArtifact;
  valueTriggerDefinition?: TTriggerDefinition;
  valueUsageContext?: TUsageContext;
  valueDosage?: TDosage;
};